move files to next structure
