+++ apps/landing/app/[lang]/favicon.ico
