+++ apps/landing/components/resources/components/Resource.tsx
