+++ apps/landing/components/resources/components/ResourcePattern.tsx
