+++ apps/landing/components/resources/components/Resources.tsx
