+++ apps/landing/components/resources/components/ResourcesContainer.tsx
