+++ apps/landing/components/resources/constants/resources.const.ts
