+++ apps/landing/components/resources/types/resource.type.ts
