+++ apps/landing/components/shared/counter/types/counter.type.ts
