+++ apps/landing/components/sponsors/Sponsors.tsx
